# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "My First API"
        }
    ],
    "paths": {
        "/api/resource.json": {
            "get": {
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "postAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create yet another post",
                "operationId": "postCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "PostStoreRequest": {
                "title": "Post storring request",
                "description": "Some simple request create post",
                "properties": {
                    "title": {
                        "title": "Title",
                        "description": "Title for storring",
                        "type": "string",
                        "format": "string",
                        "example": "random"
                    },
                    "text": {
                        "title": "Text",
                        "description": "Text for storring",
                        "type": "string",
                        "format": "text",
                        "example": "awesome"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Posts",
            "description": "Some example posts"
        }
    ],
    "security": [
        []
    ]
}